feet:changes implement

+++ home/views.py
@@ -25,7 +25,7 @@ def delete_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
 
     # Check if the user is the author of the post
-    if request.method == 'POST':
+    if request.method == 'POST' and request.user.is_authenticated:
         post.delete()
 
     return redirect('/post/')  # Replace with the URL name for your all posts page
@@ -42,12 +42,6 @@ def like_post(request, post_id):
             messages.success(request, 'You have successfully liked the post.')
     else:
         messages.warning(request, 'Please log in to like the post.')
-        # Create a new like instance
-        # like = Like(user=request.user, post=post)
-        # like.save()
-        # messages.success(request, 'You have successfully liked the post.')
-
-    # Refresh the post object to get the updated like count
     post.refresh_from_db()
 
     return redirect('/post/', post_id=post_id)  # Replace 'your_redirect_url' with the appropriate URL
@@ -68,18 +62,20 @@ def delete_comment(request, comment_id):
         # Redirect to the post detail page after deleting the comment
         return redirect('post')
     else:
-        messages.info(request, 'You Need to Login First')
+        messages.error(request, 'You Need to Login First')
         return redirect('/login/')
 
 def comment_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
 
-    if request.method == 'POST':
+    if request.method == 'POST' and request.user.is_authenticated:
         comment_text = request.POST.get('comment')
 
         # Create a new comment instance
         comment = Comment(user=request.user, post=post, text=comment_text)
         comment.save()
+    else:
+        messages.warning(request, 'For Comment this post You need to Login First')
 
     return redirect('/post/')  # Replace 'your_redirect_url' with the appropriate URL
 
@@ -90,7 +86,7 @@ def login(request):
         upass = request.POST['password']
         user = authenticate(username=uname, password=upass)
         if user is None:
-            messages.info(request, 'Username and Password are not Correct')
+            messages.error(request, 'Username and Password are not Correct')
             return redirect('/login/')
         else:
             authorize(request, user)
@@ -104,7 +100,7 @@ def login(request):
 def logout(request):
     if request.user.is_authenticated:
         deauth(request)
-        messages.info(request, 'You have Successfully Logout')
+        messages.success(request, 'You have Successfully Logout')
     return redirect('/login/')
 
 def register(request):
@@ -118,7 +114,7 @@ def register(request):
     return render(request, 'yourapp/register.html/', {'title': 'User Register', 'form': form})
 
 def post(request):
-    if request.method == 'POST':
+    if request.method == 'POST' and request.user.is_authenticated:
         # Save data to the database
         name = request.POST['name']
         desc = request.POST['desc']
@@ -126,6 +122,9 @@ def post(request):
         post = Post(name=name, desc=desc, image=image)
         post.save()
         return redirect('post')  # Redirect to the same page to display all posts
+    else:
+        if not request.user.is_authenticated:
+            messages.error(request, 'You need to Login First')
 
     # Retrieve all posts from the database
     posts = Post.objects.all().order_by('-created_at')
